requirements.txt
# Core
sqlalchemy==2.0.25
aiosqlite==0.19.0
greenlet>=3.0.0
pydantic-settings==2.1.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# YouTube
yt-dlp>=2025.6.9
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Network/Upload
httpx==0.27.0
smbprotocol==1.13.0

# Utils
pytz==2024.1
